+++ src/css/furniture-details-modal.css
@@ -25,8 +25,9 @@
     height: 100%;
     background: rgba(0, 0, 0, 0.25);
     opacity: 0;
+    padding: 100px 0;
     visibility: hidden;
-    overflow: hidden;
+    overflow: auto;
     pointer-events: none;
     transition: opacity 250ms cubic-bezier(0.4, 0, 0.2, 1),
         visibility 250ms cubic-bezier(0.4, 0, 0.2, 1);
@@ -42,13 +43,12 @@
 .modal-container {
     max-width: 335px;
     width: 100%;
-    height: 90%;
+    height: auto;
 
     position: relative;
     border-radius: 8px;
     padding: 64px 20px 44px 20px;
     margin: auto;
-    overflow-y: auto;
     background: var(--color-scheme-1-background);
     font-family: var(--font-family);
     font-weight: 400;
@@ -101,7 +101,7 @@
 
 .small-img img {
     width: 139px;
-    height: 197px;
+    height: auto;
 }
 
 .modal-content {
@@ -233,7 +233,6 @@
 
     .small-img img {
         width: 312px;
-        height: 197px;
     }
 
     .title-modal-product {
@@ -252,7 +251,6 @@
         max-width: 1090px;
         height: 649px;
         padding: 64px;
-        overflow-y: hidden;
         display: flex;
         align-items: center;
         justify-content: center;
@@ -268,10 +266,10 @@
         width: 358px;
     }
 
-    /* .img-product {
+    .img-product {
         width: 580px;
         height: 521px;
-    } */
+    }
 
     .large-img {
         width: 580px;
